Episode 7

+++ src/Commands/clear.js
@@ -5,6 +5,13 @@ const Command = require("../Structures/Command");
 module.exports = new Command({
 	name: "clear",
 	description: "Clear an amount of messages",
+	type: "BOTH",
+	slashCommandOptions: [{
+		name: "amount",
+		description: "The amount of messages to clear",
+		type: "INTEGER",
+		required: true
+	}],
 	permission: "MANAGE_MESSAGES",
 	async run(message, args, client) {
 		const amount = args[1];
@@ -23,10 +30,12 @@ module.exports = new Command({
 
 		message.channel.bulkDelete(amountParsed);
 
-		const msg = await message.channel.send(
+		const msg = await message.reply(
 			`Cleared ${amountParsed} messages!`
 		);
 
-		setTimeout(() => msg.delete(), 5000);
+		
+
+		if (msg) setTimeout(() => msg.delete(), 5000);
 	}
 });
